requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
beautifulsoup4==4.12.2
yfinance==0.2.18
openai==1.3.8
python-dotenv==1.0.0
scikit-learn==1.3.0
joblib==1.3.2
schedule==1.2.0
python-telegram-bot==20.7
lxml==4.9.3
html5lib==1.1